backend/requirements.txt
fastapi
uvicorn
requests
pdfplumber
python-docx
python-multipart